+++ binsrc/genie/helpers.py
@@ -3,6 +3,7 @@
 import os
 import shutil
 import subprocess
+import sys
 
 import nsenter
 import psutil
@@ -48,3 +49,26 @@ def get_unshare_path():
 def get_wsl_distro_name():
     """Get the WSL distribution name."""
     return os.environ['WSL_DISTRO_NAME']
+
+
+def prelaunch_checks():
+    """Check that we are on the correct platform, and as the correct user."""
+
+    # Is this Linux?
+    if not sys.platform.startswith('linux'):
+        sys.exit("genie: not executing on the Linux platform - how did we get here?")
+
+    # Is this WSL 1?
+    root_type = list(filter(lambda x: x.mountpoint == '/',
+                            psutil.disk_partitions(all=True)))[0].fstype
+    if root_type == 'lxfs' or root_type == 'wslfs':
+        sys.exit("genie: systemd is not supported under WSL 1.")
+
+    # Is this WSL 2?
+    if not os.path.exists('/run/WSL'):
+        if 'microsoft' not in os.uname():
+            sys.exit("genie: not executing under WSL 2 - how did we get here?")
+
+    # Are we effectively root?
+    if os.geteuid() != 0:
+        sys.exit("genie: must execute as root - has the setuid bit gone astray?")
